pyproject.toml
[project]
name = "viam-sdk"
description = "Viam Robotics Python SDK"
authors = [
	{name = "Naveed Jooma", email = "naveed@viam.com" }
]
license = "Apache-2.0"
readme = "README.md"
requires-python = ">=3.8.1"
dynamic = [
	'version',
]
dependencies = [
    "googleapis-common-protos>=1.65.0",
    "grpclib>=0.4.7",
    "protobuf==5.29.2",
    "typing-extensions>=4.12.2",
    "pymongo>=4.10.1",
]

[project.urls]
Homepage = "https://www.viam.com"
Documentation = "https://python.viam.dev"
Repository = "https://github.com/viamrobotics/viam-python-sdk"

[project.optional-dependencies]
mlmodel = [
	"numpy"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "src/viam/version_metadata.py"

[tool.hatch.build.targets.wheel]
packages = ["src/viam"]
artifacts = ["src/viam/rpc/libviam_rust_utils.*"]

[tool.uv]
dev-dependencies = [
	"coverage>=7.6.1",
	"google-genai>=1.20.0; python_version>='3.9'",
	"mypy-protobuf>=3.6.0",
	"myst-nb<1.0.0; python_version<'3.9'",
	"myst-nb>=1.0.0; python_version>='3.9'",
	"nbmake>=1.5.4",
	"numpy<1.25.0; python_version<'3.9'",
	"numpy>=1.26.2; python_version>='3.9'",
	"pillow>=10.4.0",
	"pyright>=1.1.382.post1",
	"pytest-asyncio>=0.24.0",
	"pytest-mock>=3.14.0",
	"pytest>=8.3.3",
	"ruff>=0.6.8",
	"sphinx-autoapi<3.0.0; python_version<'3.9'",
	"sphinx-autoapi>=3.0.0; python_version>='3.9'",
	"sphinx-rtd-theme>=2.0.0",
	"types-pillow>=10.2.0.20240822",
]

[tool.pytest.ini_options]
addopts = "-ra"
testpaths = "tests"
asyncio_mode = "auto"

[tool.coverage.run]
omit = [ "*/gen/*", "*/proto/*" ]

[tool.pyright]
include = [ "src" ]
exclude = [ "**/gen", "**/proto" ]

[tool.ruff]
line-length = 140
exclude = [
    ".direnv",
    ".git",
    ".git-rewrite",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".pytest_cache",
    ".ruff_cache",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "build",
    "dist",
    "venv",
	"gen",
	"*_grpc.py",
	"*_pb2.py",
	"*.pyi",
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
